backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.121.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.7.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.8.0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
